refactor: admin audit 도메인 응집 (domain·repository 고아 패키지를 audit 로) (#562)
- admin 하위는 feature 단위(access·announcement·template)로 응집돼 있는데 audit 만 audit/(action·service) + domain/(AdminAuditLog) + repository/(AdminAuditLogRepository) 세 곳에 흩어져 있던 것을 audit/ 한 곳으로 모음
- AdminAuditLog·AdminAuditLogRepository 를 admin/audit 로 이동하고, 같은 패키지가 된 AdminAuditService·repository 의 import 를 제거. 동작·로직 불변
- 비게 된 admin/domain·admin/repository 고아 패키지 제거
