+++ src/main/kotlin/com/depromeet/piki/admin/audit/AdminAuditLogRepository.kt
@@ -1,6 +1,5 @@
-package com.depromeet.piki.admin.repository
+package com.depromeet.piki.admin.audit
 
-import com.depromeet.piki.admin.domain.AdminAuditLog
 import org.springframework.data.jpa.repository.JpaRepository
 
 // 감사 기록 저장소. append-only — 기록만 쌓고 수정/삭제하지 않는다.
